rename fileutils to fileutil to be consistent
- all other util packages are singular, fileutils should be fileutil
